VLCKit: url support for VLCMedia
The class VLCMedia used only the trailing path part of the URL it was handed. I changed the roles of initWithPath: and initWithURL: by letting initWithPath: create an NSURL, and initWithURL: hand the complete URL to libvlc_media_new. With this change I'm able to play network streams, not just local movie files.

Signed-off-by: Derk-Jan Hartman <hartman@videolan.org>