VLCKit: Reset time, remaining time and position of VLCMediaPlayer when media is changed
When the VLCMediaPlayer switches between two media, there is a short period of time when the new media is set but when the time, remaining time and position are still set to the last time and position of the previous media. This is visible on tvOS where there is a short time when switching media where the new title is already being displayed but the old time and position is still being shown. This commit resets the time, remaining time and position of the VLCMediaPlayer when its media has changed.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>