VLCMedia: Remove media parsing in tracks information getter
A media can be parsed by a media player whitout having the VLCMediaParsedStatusInit status.
This caused the tracksInformation() method to return an empty array when a media
was parsed by a media player.

Furthermore, the user should be responsible for the parsing of the media and not
the library itself.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>