buildMobileVLCKit.sh: Add VLC_PATH variable to add additional paths
This shall be used only if you are sure what is inside those extra
paths. Please pay careful attention.  This variable shall not be set by
the normal user, and the normal $PATH variable shall not pollute this
script, as everything needed is contained in extras/tools and contribs.

Similar to f98e502fe6d15815aebec8433081425e2329baac in vlc.git.

Signed-off-by: Carola Nitz <nitz.carola@googlemail.com>