transcode: fix transcode with passthrough video
Fix transcode pipeline when using passthrough for the video track, for
example when using chromecast which doesn't try to convert the video
track.

Thus, it also fixes the playback of correct h264 stream on chromecast.

This was a regression from 553a5da7fe4f63066536ad5522bb0cf754ba8f1b
due to patch transcode-add-support-for-multiple-venc-parameters.