Headers: export class
When setting visibility to hidden by default and compiling into a
dynamic library/framework, VLCKit classes are not exposed.

Adding OBJC_VISIBLE is like using the attribute visibility("default")
in C-based languages and it signals to the compiler that it must
exposes the symbol in the final object.

See https://gcc.gnu.org/wiki/Visibility for detailed information on
symbol visibility.