VLCMediaPlayer: bump libVLC for TitleChanged event
libvlc_MediaPlayerTitleChanged event has been removed in favor of
libvlc_MediaPlayerTitleListChanged which indicates the newly available
titles and libvlc_MediaPlayerTitleSelectionChanged which indicates that
a new title has been selected by the player.

This commit bumps the tested hash to the libvlc commit adding this new
API and update VLCKit event dispatching to expose the title changed
event, but not the title list changed event since it doesn't break.