libvlc: Backport !1511
This backports videolan/vlc!1511 to update libdsm contrib (and switch to
meson), and use the new API to interrupt the I/O, easily pluggable with
the vlc interrupt API.

This fixes the input_thread_t blockage when using libdsm and when having
networks issues.

This backport was done manually to VLCKit as there are currently no
plans to do it on the mainline branch.