VLCMediaPlayer: add mediaPlayerTrack event API
Add mediaPlayerTrackAdded/Updated/Removed API to signal changes reported
from the libvlc_media_player_t object. For now, it mimics the API, by
just transforming raw char strings into NSString, but it should be
improved to report VLCMediaTrack instead.

The blocker for reporting VLCMediaTrack is that libvlc_media_track_t are
not reported by the media player as events currently, so it's not
possible to get a matching libvlc object, which makes it impossible to
get the matching VLCKit object without duplicating them.