compileAndBuildVLCKit: ignore lipo without sources
Fix the build when -a is used to select an arch that is not buildable
on either simulator or device, but available on the other, which led
lipo to complain that no files were given as input.

Since the previous archives are removed, and probed, it doesn't impact
the generation of the framework.

Example: ./compileAndBuildVLCKit.sh -ftnva aarch64