compileAndBuildVLCKit: refactor lipo construction
Refactor the checks to create the lipo-ed build for each target from
each target/arch couple behind a single function. Because of the
discrepancy between VLCKit buildsystem (expecting ${os_style}-simulator-
or ${os_style}-device- prefix) and the way VLC defines the target (which
is directly ${os_style}simulator or ${os_style}os) and the additional
discrepancy against macosx (which is not -os suffixed), some additional
code is made to make them match correctly by checking whether we run in
simulator or for device.