Set a default events handler configuration in VLCLibrary
With this proposal VLCKit will provide a new mechanism to configure events dispatch.

Any configuration can be provided with the use of the introduction of the `VLCEventsConfiguring` protocol:
```objectivec
@protocol VLCEventsConfiguring <NSObject>

- (dispatch_queue_t _Nullable) dispatchQueue;
- (BOOL) isAsync;

@end
```
Any configuration can be set globally with `+[VLCLibrary sharedEventsConfiguration]`.

`VLCEventsDefaultConfiguration` is set when `VLCLibrary` class loads at runtime.

```objectivec
@implementation VLCEventsDefaultConfiguration

- (dispatch_queue_t _Nullable)dispatchQueue {
    return nil;
}

- (BOOL)isAsync {
    return NO;
}

@end
```
The above implementation tells the events to be dispatched synchronously from any callback's calling threads.

Another `VLCEventsLegacyConfiguration` is available if any client needs to use the former events dispatching mechanism.

```objectivec
@implementation VLCEventsLegacyConfiguration

- (dispatch_queue_t _Nullable)dispatchQueue {
    return dispatch_get_main_queue();
}

- (BOOL)isAsync {
    return YES;
}

@end
```
The above implementation tells the events to be dispatched asynchronously on the global main thread's queue.
This one would be of great help when we would like to migrate to VLCKit 4 when we're already using VLCKit 3.

The configuration will be used in `-[VLCEventsHandler handleEvent:]` implementation :

```objectivec
- (void)handleEvent:(void (^)(id))handle {
    ///[truncated]
    dispatch_block_t block = ^{handle(object);};
    if (_configuration.dispatchQueue) {
        if (_configuration.isAsync)
            dispatch_async(_configuration.dispatchQueue, block);
        else
            dispatch_sync(_configuration.dispatchQueue, block);
    } else {
        block();
    }
    ///[truncated]
}
```