libvlc: use thread_local instead of a work-around in pthread
The previously used workaround that was needed for iOS 8 and the 32bit
Intel iOS simulator caused low level and hard to reproduce threading
issues.

This MR restricts the use of the work-around to the 32bit Intel
simulator and deploys thread_local everywhere else.

This raises the deployment target from iOS 8.4 to 9.0 matching the
official VLC-iOS version.