VLCMediaPlayer: provide the state in delegate
When the player is changing state, the delegate doesn't care about
receiving an NSNotification or knowing from which player it was
triggered, but it will need to know the new state from the player.

It prevents calling libvlc methods from the callback, which is usually
forbidden and works only because of the dispatch_async.