VLCMediaPlayer: Conform to new APIs to handle PIP seek properly
Picture in Picture in libvlc now require to dispatch an event when
seeking is finished and needs to know if the played media is seekable.

An API change in VLCMediaPlayer allows VLCKit users to pass a
dispatch block that will be called when seeking is complete.

Picture in Picture example is also updated to work with these changes.